requirements.txt
anyio==3.6.1
click~=7.0
fastapi~=0.68.2
h11==0.14.0
pydantic==1.10.2
sniffio==1.3.0
starlette~=0.14.2
uvicorn~=0.15.0
pip~=22.3
aioredis~=2.0.1
databases~=0.6.1
python-dotenv~=0.21.0
keyring~=18.0.1
requests~=2.22.0
asgiref~=3.5.2
PyYAML~=5.3.1
Jinja2~=3.1.2
setuptools~=45.2.0
SQLAlchemy~=1.4.41
sqlmodel~=0.0.8
alembic~=1.8.1
asyncpg~=0.26.0